refactor: 消除全部类型检查错误（29 → 0），并落地 pyright 配置
背景：项目从未把 pyright 接进任何流程（无配置、无 CI），错误无限累积到 35 条，
于是被整体当作"噪音"不再查看——**真问题就此被淹没**。逐条 triage 后：
真 bug 0 条（详见下）、纯噪音 28 条、隐式契约 1 条（未改，见末尾）。

**新配置 pyrightconfig.json**
- `extraPaths: [".", "scripts"]`：测试在函数体内 `sys.path.insert(scripts/)` 后
  `import seed_rules`，Pyright 官方明确不静态识别 sys.path 操纵（#1729），故
  只能靠 extraPaths —— 6 条 reportMissingImports 由此消除

**消除"类型偷懒"（根因，而非逐条糊告警）**
- `paddle_api.create_ocr_client() -> object` → 新增 `OCRClient` Protocol
  （docstring 早已写清协议；`progress_cb` 仅部分后端支持，故不入协议，
  调用方用 setattr 注入）。下游 import_routes 两处错误随之消失
- `get_model` / `get_reranker` 的三态哨兵（None/实例/False）使 `_model` 沦为
  object。**保持内部表示不变**，仅在出口 `cast` 表达对外契约
  `X | None`——首版改成两个变量（_load_failed）会改变可观测行为并打破 4 个
  既有测试，已回退
- `lexicon.expand._items` 的 `tuple[str, object]` → 两个 NamedTuple + isinstance
  分派。**不能用 `(kind, payload)` 可辨识联合**：实测 Pyright 不做元组跨元素相关
  窄化（`if kind == "group"` 不会窄化 payload），这与 validate_row 那次是同一个坑
- PyMuPDF stub 把 `get_text()` 重载合并成 `str|list|dict` → 出口 `cast(str, ...)`
- `_match_for(keyword: str|None)`：函数内部本就处理 None，是入参标注过窄
- `rrf_fusion(k: int=60)` → `float`：函数体只做 `1/(k+rank)`，且测试有意传 0.5
  （注解不准，改注解而非改测试）

**测试侧**
- 9 处 `list[int|None]`：`insert_pending` 返回 None 表示"该键已有行、未新建"，
  失败模式是**静默零工作量**（实测 set_status([None]) 返回 0、decide_scope 空结果）。
  在依赖"首次插入"的 8 个赋值点补断言 —— 与同文件既有写法（`assert a is not None`）一致
- test_seed_rules 删死变量 max_dim（2 条告警随之消失，无行为变化）
- test_models 加短路断言；test_api_client 用真实 httpx.Request 替代 request=None
  （忠实复刻 raise_for_status 的产物，消除替身与现实的偏差）

**未改（记录备查）**：`insert_pending() -> int | None` 把"正常去重"与
"IntegrityError 竞态"两种语义混进同一个 None，调用方无法区分（生产侧已有三种
处理方式）。改动它需单独立项评审。

验证：`pyright` 全仓 **0 errors, 0 warnings**；全量 **803 passed**。

+++ app/search/rrf.py
@@ -7,7 +7,7 @@
 
 
 def rrf_fusion(sql_results: list[dict], vector_results: list[dict],
-               k: int = 60) -> list[dict]:
+               k: float = 60) -> list[dict]:
     """融合两路有序结果，返回按 RRF 分数降序的合并 dict 列表。
 
     - sql_results：已按字段优先级有序（clause_no 精确 > title > content）
